Update Grammar
Just some house keeping, Great Project!

1)  Run on sentence, needed some finality

2) This sentence does not use commas to the best of its ability.  It conflicts with parenthetical element form.  Therefore it was reworded.

Grammar Rule in question:
Use a comma to set off parenthetical elements, as in "The Founders Bridge, which spans the Connecticut River, is falling down." By "parenthetical element," we mean a part of a sentence that can be removed without changing the essential meaning of that sentence. The parenthetical element is sometimes called "added information." This is the most difficult rule in punctuation because it is sometimes unclear what is "added" or "parenthetical" and what is essential to the meaning of a sentence.